Moved certain files into their respective folders for better separation of concerns & included bundle into gitignore file
