+++ client/src/components/RelatedProducts/relatedList/comparisonModal.jsx
